u16 instead of references to VoxelData

+++ src/app/utils/terrain/voxel/voxel_data.rs
@@ -4,7 +4,7 @@
 #[derive(Debug, PartialEq)]
 pub struct VoxelData {
 	pub name: &'static str,
-	pub id: u32,
+	pub id: u16,
 
 	pub textures: TextureSides
 }
